rename `oneClickAppsv2.ts` to `oneClickApps.ts`
